add an Image Collection filtered to Indian Point
This commit uses the long/lat of Indian Point's coast (found using the
EE Inspector tool) to filter the Landsat 8 image collection to only
scenes that cross the point.

Note that ee.Geometry.Point turns the long/lat into a vector point
feature for the filter.  The map will have Layer1 and Layer2.  Will be
good for the tutorial to show both, and increase transparency on Layer1
to show how the filter is working.